refactor(capabilities): move docs mount into platform_management capability
Move the docs virtual mount and system prompt from the standalone
platform_docs capability directly into PlatformManagementCapability.
The docs are part of the platform management surface, not a separate
concern. Delete the now-unnecessary platform_docs.rs file and remove
it from the capability registry and platform-chat harness.
