+++ crates/server/src/harnesses/platform_chat.rs
@@ -15,10 +15,7 @@ pub fn definition() -> BuiltInHarnessDefinition {
     .with_parent_name("generic")
     .with_tags(["chat", "built-in"])
     .with_roles([BuiltInHarnessRole::Chat])
-    .with_capabilities([
-        BuiltInCapabilityDefinition::new("platform_management"),
-        BuiltInCapabilityDefinition::new("platform_docs"),
-    ])
+    .with_capabilities([BuiltInCapabilityDefinition::new("platform_management")])
 }
 
 const SYSTEM_PROMPT: &str = "\
